auth-backend/requirements.txt
flask==3.1.0
python-jose[cryptography]==3.3.0
requests==2.32.3
gunicorn==23.0.0
